Change the file structure of settings page and client side api
