+++ apps/admin/src/pages/Users/index.ts
@@ -1,0 +1,3 @@
+import Users from './Users';
+
+export default Users;
