+++ apps/admin/src/pages/Users/queryGetUsers.graphql
@@ -1,0 +1,17 @@
+query GetUser {
+  getUsers {
+    id
+    avatar
+    createdAt
+    email
+    role
+    website
+    name
+    location
+    github
+    projects {
+      preview
+    }
+    banned
+  }
+}
